Okay, this is the code I'm using live right now. First off, this also has some other additions in it.
1) On the output, I did not like the hard-coded bold tags and
linebreaks, so I'm using spans. I also removed the inline styles.

2) I also added a parameter for a custom separator. Now I can use an
em dash instead of a colon between event and time.

I'm calling the function like this:
  afdn_countdownTimer('echo',3,' &mdash; ');

Each event displays like this (hopefully the em dash comes out):
  Event Name — 2 days, 5 hours, 48 minutes

****You can ignore everything above if you wish. The important part is below.

Aside from those changes, the solution to the eventlimit problem is in
lines 318-336. After a quick validation of $maxitems (like your
$eventLimit), we start the for loop. We make only one call to
cdt_format, and if the result is not null we add it to the output
variable and increment the shownEvents variable, which is checked at
the beginning of the loop. Match or exceed it, we quit.

It's probably not as long-term of a solution as what I outlined
before, but it should do the trick. Sorry, I was using 1.4 and not
trunk. If you can better implement this into 1.5, I'll certainly use
that.