Second release of alternative mailer, which finishes the plugin settings page. Plugin is now functionality complete and just calls for finishing the rest of the todo for a stable 0.9 beta release.
Swift Mailer 3.3.2 is going to be the last release for PHP4, which means that no upgrade will be required. However, Swift Mailer 4.0 is going to require a separate branch in order to maintain compatibility for PHP5 (probably just release as Alternative Mailer 2).

Deleted wp_mail.php, because it wasn't supposed to be part of the release is not used any longer.