Uploading 0.4.1
Actually fixed the critical flaw in the preg_match used to check the url highlighted by mrgreen. The fix in 0.4 didn't work full as you could still add the full url of wp-login.php as a variable and bypass the check. The preg-match now uses parse_url to only check only the path of the url and nothing else.