Uploading version 0.5.1
Fixed a bug where redirecting to a specific page was causing an endless redirection loop.

Simplified redirection logic and made it simpler. Using template_redirect no longer requires the plugin to exclude wp-login.php, wp-register.php, xmlrpc.php or anywhere in wp-admin from being inaccessible, or to check if page is a 404.

Added wp-feed.php to the list of files in the function that restricts access to feeds.

Removed sprintf from the variable that gets the current URL.