Uploading and Releasing version 0.6.5
Added the option of requiring Feed Keys even if the user is logged in.
Change the way Feed Key errors are display to the user. They no longer are displayed as a WordPress Error, Members Only now creates an RSS feed with the error in it.
Fixed a bug where redirection to wp-login.php caused a redirect loop when WordPress was installed in different folder to the site URL.
Fixed some admin page style issues with versions previous to 2.5