Tow imprvements, one bug fix:
i1: Added a "Delete selected" button to the top of the list.

i2: Added a "Select all" checkbox.

b1: Found a fix for the "clipping" problem where the matching comment ID list was getting concatenated at 1024 characters.  It's now set to be 8192 chars, which is plenty for proving the issue, but it should either be configurable, or preferably, the client should notice if the max-length is reached and adjust the max-length to accommodate the result.