Updates based on stuarts input.
Hi, just a note to people actually using this with WP 2.5+, with regards to which modules to leave off:

“Directory Protection Module” flat-out disables the /wp-admin/ address, which you need to login to your admin. I can’t see how you would be able login to your own admin with this on (wp-login.php actually forwards to /wp-admin/).

“Protect wp-content” breaks many plugins, especially AJAX-enabling plugins, which request php files from their own directories.

“Common Exploits” denies access from the post editor to the media uploader. Which is to say, when you click “Add Image” you get 301 Forbidden.

All other modules are OK to enable.

To sum up, this is a fantastic plugin and a no-exceptions required install un any website I build, but I really have to scratch my head at the inclusion of some of these modules, especially as a newbie to WP will naturally assume enabling all modules is the best way to go.