Uploading version 0.6.6
Added redirection option for users who login directly to `wp-login.php`. You can now choose to redirect them to the Front Page or to the Dashboard as normal.

Added the ability to grant a one-time view to your blog from an Administrator's IP address allowing XML RPC applications, such as [WordPress for iPhone](http://iphone.wordpress.org/) to be able to login. 

Added the ability for Administrators to remove a User's Feed Key as well as reset it.

Fixed a problem with Feed Keys not working in all situations due to a typo.