=== zd-dugg === Contributors: timglenn, justin silverton, chris black Donate link: https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=158137 Tags: digg, widget Requires at least: 2.0.2 Tested up to: 2.6.2 Stable tag: 1.0
== Description ==

A sidebar widget that allows you to list topics which you have dugg, a list of popular diggs, and a list of upcoming diggs on digg.com 
 
Thanks to Justin for building this quick and easy digg API class. 
 *http://www.jaslabs.com
 *Justin Silverton
 *justin@jaslabs.com

Thanks to Chris for starting off this digg widget using the javascript method.
 *Chris Black
 *http://cjbonline.org



== Installation ==

1. Upload `zd-dugg` folder to the `/wp-content/plugins/` directory
1. Activate the plugin through the 'Plugins' menu in WordPress
1. Add the widget to your sidebar through Design/Widget menu

== Frequently Asked Questions ==

= A question that someone might have =

An answer to that question.

= What about foo bar? =

Answer to foo bar dilemma.

== Screenshots ==

1. This screen shot description corresponds to screenshot-1.(png|jpg|jpeg|gif). Note that the screenshot is taken from
the directory of the stable readme.txt, so in this case, `/tags/4.3/screenshot-1.png` (or jpg, jpeg, gif)
2. This is the second screen shot

== Arbitrary section ==