TODO http://jonathansulo.basecamphq.com/projects/2912187/todo_items/32315071/comments
First, the frame in WP Plugin Settings for Pushit have to show entire Mobilstart web page, either by enlarge the window or via a scroll bar. Second, once one have registered for an account one should be able to close the frame and the link or button should be visiable over and under the frame.