Thijs (http://thijsvissia.nl) discovered that when WP is set as the domain's root (e.g. www.site.com) but the actual WP files are installed in a subdirectory (e.g. site.com/wp), the path used by the ajax calls to get the content is incorrect.  He found that the call to "get_bloginfo" should be using the 'wpurl' parameter instead of 'url'.
That still left the issue of the path to 'ajax-loader.gif', the animated "loading" image used as the content is being pulled. This was fixed by adding an ajax call in the javascript file to 'read-more-right-here.php' that will return, as its result, the full plugin path.

wt_rmrh.js was also heavily modified to make the above request for the path asynchronously so that if (for some reason) the path is not returned, the user can still click the "read more" link and have the content dynamically loaded (only without the loading image).  Any "read more" link clicked before that path is known will not have an associated image. Any link clicked after the path is know WILL have the image.