Major update
Added changelog to readme.
Email validation bug squashed
Cached comments now work; if user forgets to fill in antispam or makes a mistake (and the JS does not catch it) the users comment will not be lost.
Redone entire code to make it more efficient
Admin section added for changing the antispam question
Form ID and honeypot trap added to form