IE 7 and up will not allow the dynamic insert of <object> elements into the page.  The only solution (sadly) it check for those elements in the new content, and if its IE 7 or higher, re-direct to the single page display.
When <object> elements are in the new content, and CAN be shows, code was added to try and make those elements appear AFTER all text has expanded.  When collapsing, the elements are hidden BEFORE the text begins moving.  This was done because the normal "slide down/up" with those <object> elements looks really bad.