removed footer.
changed download to one file download.inc.php need is not known as of now. will remove it at later stage.

changed edit.inc.php to update and remove the error message in case a new file is created and marked for editing.

filebrowser.inc.php updated with direct link to libfile