Release of version 1.0.2
Fixed a bug that interfered with content display using I.E., Safari and Chrome on Windows OS with some themes. Caused by calls to 404 exception pages.

Removed a javascript function that was being interfered with by other plugins and replaced it with markup produced by the server
when the page is generated. Particularly noticed with Google Analystics plugin.