- New implemented feature: comment validation via JID.
It is done so that "hopefully" it should be processed after any system,
hence won't re-approve a previously disapproved comment. In particular
a comment which was filtered as "spam" won't be checked.
Also subscribed users are not checked.
Finally a malformed JID will be immediately informed. But a
non-approved comment will be trashed instead.