- The XML parser is now namespace-aware and does not do case folding.
XMPP is fully namespace-aware and as explained in XML 1.1,
section 1.2 Terminology: match [...] No case folding is performed.
These were errors from the start.

- Start of a new design: DO NOT USE THIS VERSION.

This is the very base for a brand new design which I am preparing right now.
This will be more reliable and more robust. But obviously, the current version,
though basically working, is still a mess mixing old and new design.