Updates to work around an issue where the transient might be deleted if it is being saved by the cron job while simultaneously retrieved by the front or back end. (This causes the saved transient to be deleted, which this update should prevent.)
Updated remove_cache method to only delete the transient if the embed method is not "cache". Otherwise it will set the transient to an empty array.

Updated settings_validation method to set the transient cache to the result of the get_cache method. This way a transient is always present (even if it is an empty array).