Photon update
Expands upon changes made in r622713.

If a `` isn't available but a height is, simply use the `h` API function rather than setting an artificially-large value for width and using `fit`. Now, `fit` is only used if a height and `` are available, but no explicit width could be determined.

With these changes, `resize`, `w`, and `h` are used in the majority of cases.

Props westi and mkoenig.