Thanks to mgropel
After activating the bbpressmoderation plug-in my bbpress forum topics were no longer visible. After some digging, it appears to me that there is a bug in the posts_where function. This function was not taking into account those wp sites that use a database prefix. I fixed this by making the following change in the posts_where function:
global $wpdb;
		$where = str_ireplace($wpdb->prefix."wp_posts.post_status = 'publish'", "(".$wpdb->prefix."wp_posts.post_status = 'publish' OR ".$wpdb->prefix."wp_posts.post_status = 'pending')", $where);