Add email_post_changes_admin_email_fallback filter.
By default, EPC will add the admin_email to the list out outbound addresses if the list would otherwise be empty. This runs before the email_post_changes_emails filter, though. It's possible that another plugin will add emails to the list and will not want the admin_email to be automatically included. Removing the admin_email via the email_post_changes_emails filter would require checking multiple EPC settings to determine if it was placed there automatically or by the user's configuration, so a filter is a cleaner approach to letting other plugins control the behavior.