0.9.0 Beta Improved how the cache folder is stored to make it more friendly to switching between Dev and Production environments.
Previously the entire PATH to the cache folder was stored in one
variable in the database. That resulted in a broken PATH when WordPress
was moved to a different system for example going from a dev server to
a production server.

Now the plugin stores the path in two variables, ‘cache_base’ and
‘cache_folder’. By Default ‘cache_base’ contains ‘wp-content’ which the
plugin then turns into the contents of ‘WP_CONTENT_DIR’ and is
prepended to the contents of ‘cache_folder’ which by default contains
‘/uploads/cache/‘. The idea is that by default I want the cache in the
wp-contents folder and to play nice with moving the server around while
still making it possible to set the cache outside of the wp-contents
folder in special circumstances.

I also centralized the process of building and validating the completed
cache folder path to ffg_cache::cache_folder(). The cache_folder method
will combine cache_base and cache_folder making sure it’s an accessible
writable directory.