Escape user input in Settings fields.
A malicious Administrator could enter HTML into the text fields to break out of them and execute arbitrary JavaScript. For example, "><img src=x onerror=prompt(1)>

Because this requires an Administrator account in order to exploit, it is a minor vulnerability and unlikely to be a useful attack vector.
props trizaeron
xref: https://hackerone.com/reports/9375