Partially test for logging into the XML-RPC interface with a valid application password.
This still needs to be finished, but I'm commiting it now because I'm going to take a different direction with the XML parsing, and I want to save the work that's already been done in case it's useful in the future.

After spending way too much time getting SimpleXML to do what I want, it occured to me that in the current situation, all I really need to do is use strpos() to search for the expected strings. The next commit will remove convertXmlRpcResponseToAssociativeArray() and use strpos() to finish this test, but now it will be stored in the source history in case a more complex use case occurs in the future.