Update the parameters for Google Services to return a refresh token
If the user has already authorised the application then on subsequent
authorisation flows, consent is not requested and as a result, no
refresh token is returned. The access type is set to `offline` and a
refresh token is always needed, otherwise there is no way to get a new
valid access token.

This change adds the `prompt=consent`query parameter which means that
user consent is requested each time, and the refresh token is returned.