Google Base Class: Ensure consistency of the action parameters
In the OAuth2 class, after authentication the `keyring_access_token`
filter is applied to the token but the second parameter is an array of
the token data received from the third party service. In the Google Base
class, when refreshing, this parameter is the current
`Keyring_Access_Token` object, and so causees errors on any registered
filter callback functions relying on that value.

This PR changes fixes that. The filter is not being used as part of this
codebase, but it could conceivably break sites that are using this
filter.